hotshot_utils

Module anytrace

Error utilities, intended to function as a replacement to anyhow.

Modules

macros ๐

Macros

Macros

bail

Return an error.

debug

Create an error at the debug level.

ensure

Check that the given condition holds, otherwise return an error.

error

Create an error at the error level.

info

Create an error at the info level.

line_info

Print the file and line number of the location this macro is invoked

log

Log a anytrace::Error at the corresponding level.

trace

Create an error at the trace level.

warn

Create an error at the warn level.

Structs

Error

main error type

Enums

Level

Possible log levels

Constants

DEFAULT_LOG_LEVEL

Default log level for the crate

Traits

Context

Trait for converting error types to a Result<T>.

Log

Trait for logging errors

Wrap

Trait for a std::result::Result that can be wrapped into a Result

Functions

concatenate ๐

Prepend an error to its cause

Type Aliases

Result

Alias for the main Result type used by the crate.