Crate hotshot_utils

General (not HotShot-specific) utilities

Modules

anytrace

Error utilities, intended to function as a replacement to anyhow.

Macros

bail

Return an error.

debug

Create an error at the debug level.

ensure

Check that the given condition holds, otherwise return an error.

error

Create an error at the error level.

info

Create an error at the info level.

line_info

Print the file and line number of the location this macro is invoked

log

Log a anytrace::Error at the corresponding level.

trace

Create an error at the trace level.

warn

Create an error at the warn level.